restructure: remove utils
