+++ tests/test_utils.py
@@ -1,8 +1,13 @@
-from microdata_validator import utils
-from microdata_validator.utils import ParseMetadataError
-from microdata_validator.repository import local_storage
 from pathlib import Path
+
 import pytest
+
+from microdata_validator.schema import (
+    inline_metadata_references,
+    ParseMetadataError
+)
+from microdata_validator.repository import local_storage
+
 
 RESOURCES_DIR = "tests/resources/inline_metadata"
 SIMPLE_JSON = Path(f"{RESOURCES_DIR}/simple_referenced.json")
@@ -13,7 +18,7 @@ REF_DIR = Path("tests/resources/ref_directory")
 
 
 def test_inline_simple():
-    actual_inlined = utils.inline_metadata_references(
+    actual_inlined = inline_metadata_references(
         SIMPLE_JSON, REF_DIR
     )
     assert actual_inlined == local_storage.load_json(
@@ -23,20 +28,20 @@ def test_inline_simple():
 
 def test_inline_simple_bad_ref():
     with pytest.raises(FileNotFoundError):
-        utils.inline_metadata_references(
+        inline_metadata_references(
             SIMPLE_JSON_BAD_REF, REF_DIR
         )
 
 
 def test_inline_invalid_ref_dir():
     with pytest.raises(ParseMetadataError) as e:
-        utils.inline_metadata_references(
+        inline_metadata_references(
             SIMPLE_JSON, None
         )
     assert "No supplied reference directory" in str(e)
 
     with pytest.raises(ParseMetadataError) as e:
-        utils.inline_metadata_references(
+        inline_metadata_references(
             SIMPLE_JSON, SIMPLE_JSON
         )
     assert "Supplied reference directory is invalid" in str(e)
@@ -44,6 +49,6 @@ def test_inline_invalid_ref_dir():
 
 def test_inline_invalid_metadata_file_path():
     with pytest.raises(FileNotFoundError):
-        utils.inline_metadata_references(
+        inline_metadata_references(
             NONEXISTENT_FILE_PATH, REF_DIR
         )
